Refoctor code mysql_storage.* to different files
