+++ CMakeLists.txt
@@ -21,7 +21,9 @@ list (APPEND WSJCPP_SOURCES "src/main.cpp")
 list (APPEND WSJCPP_SOURCES "./src/database/models_diff.cpp")
 list (APPEND WSJCPP_SOURCES "./src/database/models_basic.cpp")
 list (APPEND WSJCPP_SOURCES "./src/database/models_cache.cpp")
+list (APPEND WSJCPP_SOURCES "./src/database/mysql_storage_connection.cpp")
 list (APPEND WSJCPP_SOURCES "./src/database/mysql_storage.cpp")
+list (APPEND WSJCPP_SOURCES "./src/database/mysql_storage_updates.cpp")
 
 # server
 list (APPEND WSJCPP_SOURCES "./src/server/webdiff_http_server.cpp")
